Move `testIntegerObject` function to `eval-integer-expression_test` file
