+++ evaluator/test/eval-integer-expression_test.go
@@ -1,6 +1,10 @@
 package evaluator_test
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/MBATheGamer/lang_core/object"
+)
 
 type TestIntegerExpression struct {
 	input    string
@@ -17,4 +21,20 @@ func TestEvalIntegerExpression(t *testing.T) {
 		var evaluated = testEval(test.input)
 		testIntegerObject(t, evaluated, test.expected)
 	}
+}
+
+func testIntegerObject(t *testing.T, obj object.Object, expected int64) bool {
+	var result, ok = obj.(*object.Integer)
+
+	if !ok {
+		t.Errorf("object is not Integer. got=%T (%+v)", obj, obj)
+		return false
+	}
+
+	if result.Value != expected {
+		t.Errorf("object has wrong value. got=%d, want=%d", result.Value, expected)
+		return false
+	}
+
+	return true
 }
